Cleaning up this KaiserKernel default, which I'm pretty happy with. But before the big v0.5 release, we have two bugs that need to get resolved:
1. https://github.com/IntegralEquations/HMatrices.jl/issues/80
The `cholesky` method for Hermitian kernel matrices would be super
useful and save everybody time and RAM. I'd love to get this issue
resolved and this improved method landed.

-- https://github.com/ludvigak/FINUFFT.jl/issues/80
The newest `finufft_jll` package clearly has some compat-breaking bugs
with `FINUFFT.jl`. So for now, I've pinned it at the previous working
version. But that shouldn't really be a dep in the tree, so I'll
probably sit on this code until that gets resolved.